wip: splitting baserow class
