+++ libs/data/src/common/baserow/project.ts
@@ -1,0 +1,89 @@
+import { Baserow } from './baserow'
+import { RawProject } from '../../project/types'
+import { LinkObject, Project } from './types'
+import { Theme } from '../../theme/themes'
+
+export class BaserowProject extends Baserow {
+  private readonly _projectTableId = 305253
+
+  async getValidProjects(): Promise<RawProject[]> {
+    const baserowProjects = await this._getTableData<Project>(this._projectTableId)
+    const validBaserowProjects = baserowProjects.filter((value) => {
+      return value.Publié
+    })
+
+    const baserowThemes = await this._getTableData<Theme>(this._themeTableId)
+
+    const projects: RawProject[] = []
+    for (const project of validBaserowProjects) {
+      try {
+        const result = await this._convertToRawProjectType(project, baserowThemes)
+        projects.push(result)
+        console.info(`successfully loaded project ${project.id}`)
+        await this._delay(300)
+      } catch (error) {
+        console.error(`Error processing project ${project.id}:`, error)
+      }
+    }
+    return projects
+  }
+
+  private async _convertToRawProjectType(baserowProject: Project, baserawThemes: Theme[]): Promise<RawProject> {
+    const imageName = await this._handleImage(baserowProject.Image)
+
+    return {
+      id: baserowProject.id,
+      slug: baserowProject.Nom,
+      title: baserowProject.Titre,
+      nameTag: baserowProject.NameTag,
+      shortDescription: baserowProject['Description courte'],
+      image: this._imagePath + imageName,
+      longDescription: baserowProject['Qu’est-ce que c’est ?'],
+      moreDescription: baserowProject['Pour aller plus loin'],
+      themes: this._generateThemeList(baserowProject['Thématique principale'], baserowProject['Thématiques secondaires'], baserawThemes),
+      mainTheme: this._generateMainTheme(baserowProject['Thématique principale'], baserawThemes),
+      programs: this._generateProgramList(baserowProject.Dispositifs),
+      linkedProjects: this._generateLinkedProjectList(baserowProject['Projets complémentaires']),
+      priority: baserowProject.Prio
+    }
+  }
+
+  private _generateMainTheme(mainTheme: LinkObject[], baserowThemes: Theme[]): string {
+    if (mainTheme.length != 1) {
+      console.warn('Missing mainTheme Or mainTheme not unique in a field')
+    }
+    const themeId = mainTheme[0].id
+    const matchingTheme = baserowThemes.find((theme) => theme.id === themeId)
+    if (matchingTheme === undefined) {
+      console.warn('theme not found in baserow data (should not happen!)')
+      return ''
+    }
+    return matchingTheme['Nom (Tech)']
+  }
+
+  private _generateThemeList(mainTheme: LinkObject[], secondaryThemes: LinkObject[], baserowThemes: Theme[]): string[] {
+    const themeList = [this._generateMainTheme(mainTheme, baserowThemes)]
+    secondaryThemes.forEach((secondaryTheme) => {
+      const themeId = secondaryTheme.id
+      const matchingTheme = baserowThemes.find((theme) => theme.id === themeId)
+      if (matchingTheme) {
+        themeList.push(matchingTheme['Nom (Tech)'])
+      } else {
+        console.warn('theme not found in baserow data (should not happen!)')
+      }
+    })
+    return themeList
+  }
+
+  private _generateProgramList(programs: LinkObject[]): string[] {
+    return programs.map((program) => {
+      return program.value
+    })
+  }
+
+  private _generateLinkedProjectList(projects: LinkObject[]): number[] {
+    return projects.map((project) => {
+      return project.id
+    })
+  }
+}
